% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compute.gE.R
\name{compute.gE}
\alias{compute.gE}
\title{Compute just gE from latitude.}
\usage{
compute.gE(phi)
}
\arguments{
\item{phi}{Latitude in degrees North.}
}
\description{
\code{compute.gE} computes an estimate of the seasonal growth response to insolation at
each month in the year at latitude phi.
}
\details{
The monthly values returned in gE are used to scale the growth responses to temperature
and soil moisture in VS-Lite.
}

