% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/daylength.factor.from.lat.R
\name{daylength.factor.from.lat}
\alias{daylength.factor.from.lat}
\title{Compute normalized daylength.}
\usage{
daylength.factor.from.lat(phi, return.ndl.and.cdays = FALSE)
}
\arguments{
\item{phi}{Latitude in degrees.}

\item{return.ndl.and.cdays}{Logical; return auxiliary variabiles ndl and cdays? Default
is FALSE.}
}
\description{
\code{daylength.factor.from.lat} computes a normalized daylength factor for each month,
used to modulate growth responses in VS-Lite to account for seasonal variations in
insolation.
}
\details{
The small difference in calculation for leap-years is negleced.
}

