% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{eyear}
\alias{eyear}
\title{End years of test chronologies and PRISM climate datasets.}
\format{\preformatted{ num [1, 1] 1984
}}
\usage{
eyear
}
\description{
End years of test chronologies and PRISM climate datasets.
}
\keyword{datasets}

