% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/leakybucket.submonthly.R
\name{leakybucket.submonthly}
\alias{leakybucket.submonthly}
\title{Run CPC Leaky Bucket model (submonthly version).}
\usage{
leakybucket.submonthly(syear, eyear, phi, T, P, Mmax = 0.76, Mmin = 0.01,
  alph = 0.093, m.th = 4.886, mu.th = 5.8, rootd = 1000, M0 = 0.2)
}
\arguments{
\item{syear}{Start year of simulation.}

\item{eyear}{End year of simulation.}

\item{phi}{Latitude of site (in degrees N).}

\item{T}{(12 x Nyrs) Matrix of ordered mean monthly temperatures (in degEes C).}

\item{P}{(12 x Nyrs) Matrix of ordered accumulated monthly precipitation (in mm).}

\item{Mmax}{Scalar maximum soil moisture held by the soil (in v/v).}

\item{Mmin}{Scalar minimum soil moisture (for error-catching) (in v/v).}

\item{alph}{Scalar runoff parameter 1 (in inverse months).}

\item{m.th}{Scalar runoff parameter 3 (unitless).}

\item{mu.th}{Scalar runoff parameter 2 (unitless).}

\item{rootd}{Scalar root/"bucket" depth (in mm).}

\item{M0}{Initial value for previous month's soil moisture at t = 1 (in v/v).}
}
\value{
M Soil moisture computed via the CPC Leaky Bucket model (in v/v, 12 x Nyrs).

potEv Potential evapotranspiration computed via Thornthwaite's 1947 scheme (in mm).
}
\description{
\code{leakybucket.submonthly} simulates soil moisture with submonthly time step.
}
\details{
Modifications of Suz Tolwinski-Ward's monthly time-step code by Nick Graham in 2011.
Ported to R by SETW in 2015. Implementation of CPC Leaky Bucket model as described in
Huang et al., 'Analysis of Model-Calculated Soil Moisture over the United States
(1931-1993) and Applications to Long-Range Temperature Forecasts,' J. Clim. (1995)
#' Fixes to model consistent with Matlab VSLite version 2.5.
}
\seealso{
\code{\link{leakybucket.monthly}}
}

