% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{P}
\alias{P}
\title{Precipitation test data. From PRISM.}
\format{\preformatted{ num [1:12, 1:90, 1:4] 160.03 83.3 32.76 15.12 5.64 ...
}}
\source{
\url{http://www.prism.oregonstate.edu/}
}
\usage{
P
}
\description{
Precipitation test data. From PRISM.
}
\keyword{datasets}

