% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{sitecoords}
\alias{sitecoords}
\title{Coordinates of test sites.}
\format{\preformatted{ num [1:4, 1:2] 36.5 34.2 39 39.3 -118.2 ...
}}
\usage{
sitecoords
}
\description{
Coordinates of test sites.
}
\keyword{datasets}

