% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/std.ramp.R
\name{std.ramp}
\alias{std.ramp}
\title{"Standard ramp" function for building growth response functions.}
\usage{
std.ramp(x, x1, x2)
}
\arguments{
\item{x}{The value at which we want to evaluate the ramp function.}

\item{x1}{The lower bound of the support of the nonzero part of the ramp.}

\item{x2}{The lower bound of the range of the preimage of 1.}
}
\description{
\code{std.ramp.r} takes a lower and upper bounds, and creates a linear response
between 0 and 1 for values in between the bounds.  Values below (above) the lower (upper)
bound are assigned a value of zero (one).
}

