% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.R
\name{test_VSLiteR}
\alias{test_VSLiteR}
\title{Test port of VSLite to R, and draw some figures.}
\usage{
test_VSLiteR(site = 1)
}
\arguments{
\item{site}{Index 1, 2, 3, or 4, specifies which of 4 test sites to run.}
}
\description{
Test VSLiteR code.
}
\seealso{
\code{\link{VSLite}}
}

