% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{Tmp}
\alias{Tmp}
\title{Temperature test data. From PRISM.}
\format{\preformatted{ num [1:12, 1:90, 1:4] -5.72 -3.68 -2.72 -0.245 3.735 ...
}}
\source{
\url{http://www.prism.oregonstate.edu/}
}
\usage{
Tmp
}
\description{
Temperature test data. From PRISM.
}
\keyword{datasets}

