% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{trw.obs}
\alias{trw.obs}
\title{Observed dendrochronologies for running tests.}
\format{\preformatted{ num [1:90, 1:4] -0.2809 -0.192 -0.4103 -0.5963 -0.0464 ...
}}
\source{
\url{https://catalog.data.gov/dataset/international-tree-ring-data-bank-itrdb}
}
\usage{
trw.obs
}
\description{
Observed dendrochronologies for running tests.
}
\keyword{datasets}

