\name{VSLiteR-package}
\alias{VSLiteR-package}
\alias{VSLiteR}
\docType{package}
\title{VS-Lite model of tree-ring growth, in R.}
\description{Efficient simulation of dendrochronologies from climate input data; a "lite" version of the full, biology-based Vaganov-Shashkin model.}
\details{
\tabular{ll}{
Package: \tab VSLiteR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-11\cr
License: \tab MIT \cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Suz Tolwinski-Ward
<suzski@gmail.com>
}
\references{
Tolwinski-Ward, S.E., M.N. Evans, M.K. Hughes, K.J. Anchukaitis, An efficient forward model of the climate controls on interannual variation in tree-ring width, Climate Dynamics, doi:10.1007/s00382-010-0945-5, (2011). 

Tolwinski-Ward, S.E., K.J. Anchukaitis, M.N. Evans, Bayesian parameter estimation and interpretation for an intermediate model of tree-ring width , Climate of the Past, doi:10.5194/cpd-9-615-2013, (2013). 

Tolwinski-Ward, S.E., M.P. Tingley, M.N. Evans, M.K. Hughes, D.W. Nychka, Probabilistic reconstructions of local temperature and soil moisture from tree-ring data with potentially time-varying climatic response, Climate Dynamics, doi:10.1007/s00382-014-2139-z, (2014). 
}
\keyword{ package }
\examples{
}
