% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/VSLiteR.R
\name{VSLiteR}
\alias{VSLiteR}
\title{VS-Lite model of tree ring width growth.}
\usage{
VSLiteR(syear, eyear, phi, T1, T2, M1, M2, T, P, varargin, Mmax = 0.76,
  Mmin = 0.01, alph = 0.093, m.th = 4.886, mu.th = 5.8, rootd = 1000,
  M0 = 0.2, substep = 0, I_0 = 1, I_f = 12, hydroclim = "P")
}
\arguments{
\item{syear}{Start year of simulation.}

\item{eyear}{End year of simulation.}

\item{phi}{Latitude of site (in degrees N).}

\item{T1}{Lower temperature threshold for growth to begin (scalar, deg. C).}

\item{M1}{Lower moisture threshold for growth to begin (scalar, v.v).}

\item{M2}{Upper moisture threshold for growth sensitivity to moisture (scalar, v/v).}

\item{T}{(12 x Nyrs) Matrix of ordered mean monthly temperatures (in degEes C).}

\item{P}{(12 x Nyrs) Matrix of ordered accumulated monthly precipitation (in mm).}

\item{Mmax}{Scalar maximum soil moisture held by the soil (in v/v).}

\item{Mmin}{Scalar minimum soil moisture (for error-catching) (in v/v).}

\item{alph}{Scalar runoff parameter 1 (in inverse months).}

\item{m.th}{Scalar runoff parameter 3 (unitless).}

\item{mu.th}{Scalar runoff parameter 2 (unitless).}

\item{rootd}{Scalar root/"bucket" depth (in mm).}

\item{M0}{Initial value for previous month's soil moisture at t = 1 (in v/v).}

\item{I_0}{}

\item{I_f}{}

\item{hydroclim}{Switch; value is either "P" (default) or "M" depending on whether the
second input climate variable is precipitation, in which case soil moisture is estimated
using the Leaky Bucket model of the CPC, or soil moisture, in which case the inputs are
used directly to compute the growth response.}

\item{T1}{Upper temperature threshold for growth sensitivity to temp (scalar, deg. C).}

\item{subtep}{}
}
\value{
trw

gT

gM

gE

M

potEv

sample.mean.width

sample.std.width
}
\description{
\code{VSLiteR} simulates tree ring width growth.
}
\details{
R port of VS-Lite Model of Tree Ring Width by Suz TOlwinski-Ward, 2015. For more references,
see xxxxyyyyyzzzz.
}
\seealso{
\code{\link{compute.gE}},\code{\link{std.ramp}},\code{\link{leakybucket.monthly}},\code{\link{leakybucket.submonthly}}
}

